+++ Krowi_AchievementFilter/Localization/GlobalStrings.lua
@@ -29,6 +29,7 @@ function globalStrings.Load(L)
     L["Options"] = GAMEOPTIONS_MENU;
     L["General"] = GENERAL;
     L["Key Binding"] = KEY_BINDING;
+    L["Missing"] = ADDON_MISSING;
 
     -- Options
     L["Info"] = INFO;